refactor: recurrence feature is in its own module
