+++ src/environments/environment.development.ts
@@ -3,6 +3,6 @@ export const ENV: { auth: { clientId: string }; api: { maps: string } } = {
     clientId: '4efnrpgm45mbhp7l0lhokfb7v0'
   },
   api: {
-    maps: 'AIzaSyC3IxN7zlutIg-JmyGbYOyaEh6vbUc7ZWM'
+    maps: ''
   }
 };
