+++ src/features/place/components/place-results-dropdown/place-results-dropdown.component.ts
