+++ src/features/place/index.ts
@@ -1,0 +1,3 @@
+export * from './modules/place.feature.module';
+export * from './providers';
+export * from './definitions';
