+++ src/features/place/providers/queries/search-place.query.provider.ts
@@ -1,10 +1,10 @@
 import { Observable } from 'rxjs';
 import { FactoryProvider } from '@angular/core';
-import { Place } from '@features/places/definitions/places';
+import { Place } from '@features/place/definitions/places';
 
 export type SearchPlaceQuery = (search: string) => Observable<Place[]>;
 
-export const SEARCH_PLACE_QUERY: symbol = Symbol('places.search-place.query');
+export const SEARCH_PLACE_QUERY: symbol = Symbol('place.search-place.query');
 
 export const searchPlaceQueryProvider = <TDependencies>(
   useFactory: (...providers: never[]) => SearchPlaceQuery,
