+++ src/features/planning/actions/index.ts
@@ -1,3 +1,2 @@
 export * from './schedule-fare/schedule-fare.action';
 export * from './affect-return/affect-return.action';
-export * from './predict-recurrence/predict-recurrence.action';
