--- src/features/planning/components/modal/experimental-modal.component.css
@@ -1,16 +1,0 @@
-.modal {
-  position: fixed;
-  inset: 0;
-  background: rgb(0 0 0 / 70%);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal-content {
-  background: white;
-  padding: 1em;
-  width: 80%;
-  max-width: 500px;
-  border-radius: 4px;
-}
