--- src/features/planning/components/modal/experimental-modal.component.html
@@ -1,6 +1,0 @@
-<div class="modal">
-  <div class="modal-content">
-    <ng-content></ng-content>
-    <button (click)="close()">Fermer la modale</button>
-  </div>
-</div>
