--- src/features/planning/components/modal/experimental-modal.component.ts
@@ -1,14 +1,0 @@
-import { Component, EventEmitter, Output } from '@angular/core';
-
-@Component({
-  selector: 'app-experimental-modal',
-  templateUrl: './experimental-modal.component.html',
-  styleUrls: ['./experimental-modal.component.css']
-})
-export class ExperimentalModalComponent {
-  @Output() public closeModal: EventEmitter<void> = new EventEmitter<void>();
-
-  public close(): void {
-    this.closeModal.emit();
-  }
-}
