+++ src/features/planning/providers/actions/index.ts
@@ -1,3 +1,2 @@
 export * from './schedule-fare.action.provider';
 export * from './affect-return.action.provider';
-export * from './predict-recurrence.action.provider';
