+++ src/features/recurrence/components/index.ts
@@ -1,0 +1,4 @@
+import { PredictRecurrenceComponent } from './predict-recurrence/predict-recurrence.component';
+
+// eslint-disable-next-line @typescript-eslint/typedef
+export const COMPONENTS = [PredictRecurrenceComponent];
