+++ src/features/recurrence/modules/index.ts
@@ -1,0 +1,1 @@
+export * from './recurrence.feature.module';
